requirements.txt
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
google-auth==2.40.2
google-auth-oauthlib==1.2.2
greenlet==3.2.2
gspread==6.2.1
idna==3.10
oauthlib==3.2.2
playwright==1.52.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyee==13.0.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
typing_extensions==4.13.2
urllib3==2.4.0
